Fixed pressure plate entity param, peverified
used peverify to locate any other injection related problems, and
corrected them

+++ OTAPI.Modifications/OTAPI.Modifications.Npc.Strike/Callbacks/Strike.cs
@@ -9,7 +9,7 @@
 		internal static bool Strike
 		(
 			global::Terraria.NPC npc,
-			ref int cancelResult,
+			ref double cancelResult,
 			ref int Damage,
 			ref float knockBack,
 			ref int hitDirection,
