+++ OTAPI.Modifications/OTAPI.Modifications.Npc.Strike/Hooks/Strike.cs
@@ -8,7 +8,7 @@
             public delegate HookResult StrikeHandler
             (
                 global::Terraria.NPC npc,
-                ref int cancelResult,
+                ref double cancelResult,
                 ref int Damage,
                 ref float knockBack,
                 ref int hitDirection,
